Rename miplib package to a more general name for loading
