+++ README.md
@@ -91,7 +91,7 @@ might require, it works the exact same way it just doesn't stop after `n` instan
 [MIPLIB](https://miplib.zib.de/) 2017 instances can be loaded into a PySCIPOpt model using the `Loader` class.
 
 ```python
-from geco.mips.miplib.base import Loader
+from geco.mips.loading.miplib import Loader
 
 instance = Loader().load_instance('INSTANCE_NAME.mps.gz')
 ```
