+++ geco/mips/__init__.py
@@ -8,4 +8,4 @@ import geco.mips.max_cut as max_cut
 import geco.mips.production_planning as production_planning
 import geco.mips.packing as packing
 import geco.mips.graph_coloring as graph_coloring
-import geco.mips.miplib as miplib
+import geco.mips.loading as miplib
